% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstock.R
\name{bootstock}
\alias{bootstock}
\title{bootstock: Bootstrapping of PBT and GSI assignments to generate stock CIs}
\usage{
bootstock(stockAssignments = NULL, tagRates = NULL,
  reallocateTable = NULL, bootstraps = 500, ci = 0.9)
}
\arguments{
\item{stockAssignments}{description here}

\item{tagRates}{description here}

\item{reallocateTable}{description here}

\item{bootstraps}{The number of bootstraps to be performed}

\item{ci}{The confidence interval to be generated}
}
\description{
Bootstrapping of PBT and GSI assignments to generate stock CIs around point estimates
}
\author{
Mike Ackerman
}

