% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleStockData.R
\docType{data}
\name{exampleStockData}
\alias{exampleStockData}
\title{Example of Stock Assignments for resampit()}
\format{A data frame. Must contain a column named 'stock' containing the stock assignments.}
\usage{
exampleStockData
}
\description{
An example of stock assignment data to be bootstrapped using resampit() or bootstock(). This example
data is from parentage-based tagging (PBT).
}
\keyword{data}
\keyword{stock}

