% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleTagRates.R
\docType{data}
\name{exampleTagRates}
\alias{exampleTagRates}
\title{Example of Tagging Rates to be used by resampit () or bootstock()}
\format{A data frame. Any stock defined in the 'stock' column of the \code{stockAssignments} must be present in the
'stock' column in the \code{tagRates}}
\usage{
exampleTagRates
}
\description{
An example of stock tagging rates to be bootstrapped using resampit() or bootstock(). Each stock is 'expanded'
by the tagging rates defined.
}
\keyword{data}
\keyword{rate}
\keyword{tag}

