% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampit.R
\name{resampit}
\alias{resampit}
\title{resampit: Simple bootstrapping of fish stocks to generate CIs}
\usage{
resampit(stockAssignments = NULL, tagRates = NULL, bootstraps = 500,
  ci = 0.9)
}
\arguments{
\item{stockAssignments}{description here}

\item{tagRates}{description here}

\item{bootstraps}{The number of bootstraps to be performed}

\item{ci}{The confidence interval to be generated}

\item{abundance}{If abundance results are desired, enter the estimated abundance}
}
\description{
Bootstrap stock assignments to generate CIs
}
\author{
Mike Ackerman
}

